feat(web): split WoW tab into Guild News and Crafting Boards
Replace the single WoW sidebar entry with two focused tabs:
- Guild News: tracker list per WoW guild/realm
- Crafting Boards: board config + orders with status filter

Remove the now-unused WowTab.vue.
